+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,4 @@
 <resources>
     <string name="app_name">Google Utilities</string>
+    <string name="google_map_key">AIzaSyAYKX_dHfYIc9Cbla7am0Hq2-A9ua2meKE</string>
 </resources>
