+++ utilities/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Google Utilities</string>
+    <string name="app_name">Google Services</string>
 </resources>
